+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## CHANGELOG
 
+### 3.3.0b1 (21-Apr-2026)
+- Iru-specific updates [Pull Request #88](https://github.com/dan-snelson/Mac-Health-Check/pull/88); thanks, @kgolden-code!
+- New `checkWiFiStrength()` check [Pull Request #88](https://github.com/dan-snelson/Mac-Health-Check/pull/88); thanks, @kgolden-code!
+
 ### 3.2.0 (02-Apr-2026)
 - Preserved user-provided local `organizationOverlayiconURL` files by downloading remote overlay icons to a per-run temporary file and only cleaning up that script-managed asset at exit (Thanks for the heads-up, @brian_b!)
 - Corrected Jamf Pro inventory warning text for non-SSO sessions so omitted `-endUsername` logging now explains that no SSO username was available for the logged-in user.